# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShareHub API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/me:
    get:
      summary: 个人中心聚合信息
      description: 需要登录后才会返回完整的用户统计，联调可通过 X-User-Key 透传用户标识；首次访问会自动为当前用户补建默认档案；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/resources:
    get:
      summary: 获取我的资料列表
      description: 当前实现按请求用户返回自己的资料工作台数据，默认不限定发布态；可通过 `status`、`visibility` 过滤草稿或已发布数据。联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；`status`、`visibility` 传空字符串或仅空白时按未传处理，带前后空白时会先裁剪再参与筛选；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: status
          in: query
          schema:
            type: string
        - name: visibility
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/roadmaps:
    get:
      summary: 获取我的路线列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；`status` 传空字符串或仅空白时按未传处理，带前后空白时会先裁剪再参与筛选；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadmapWorkbenchPageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/favorites:
    get:
      summary: 获取我的收藏资料列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/favorite-notes:
    get:
      summary: 获取我的收藏笔记列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelatedNotePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/notes:
    get:
      summary: 获取我的笔记列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`；`status` 传空字符串或仅空白时按未传处理，带前后空白时会先裁剪再参与筛选
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/note-history:
    get:
      summary: 获取我的笔记浏览历史
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelatedNotePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me/resumes:
    get:
      summary: 获取我的简历分页列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；`status`、`templateKey`、`keyword` 传空字符串或仅空白时按未传处理，带前后空白时会先裁剪再参与筛选；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: status
          in: query
          schema:
            type: string
        - name: templateKey
          in: query
          schema:
            type: string
          description: 按模板筛选
        - name: keyword
          in: query
          schema:
            type: string
          description: 按模板名或文件名模糊筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/resources:
    get:
      summary: 资料分页列表
      description: 当前公开资料列表默认只返回 `PUBLISHED`；`status` 为空字符串、仅空白，或裁剪后没有有效状态项时按未传处理，带前后空白时会先裁剪再参与筛选；`category` 当前映射到资源 `type` 字段；`sortBy=latest` 按 `updatedAt DESC`，`sortBy=hot` 按点赞数降序再按更新时间降序
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 12
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: visibility
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
          description: 当前实现映射到资源 `type` 字段
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [latest, hot]
      responses:
        '200':
          description: 成功
    post:
      summary: 创建资料
      description: 需要当前用户身份，联调可通过 `X-User-Key` 指定用户；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/roadmaps:
    get:
      summary: 路线分页列表
      description: 当前列表仍返回全量路线，`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底，且返回结果按 `id` 倒序排列
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
    post:
      summary: 创建路线
      description: 需要当前用户身份；创建记录的 owner 固定取请求用户，联调可通过 `X-User-Key` 指定用户；`status` 未传、为 `null`、空字符串或仅空白时默认补成 `PUBLISHED`，带前后空白时会先裁剪再保存；未带身份时返回 `401 NOT_LOGGED_IN`，用户已被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
...
  /api/notes:
    get:
      summary: 笔记分页列表
      description: 需要当前用户身份；当前实现只返回请求用户自己的笔记；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`；`status` 传空字符串或仅空白时按未传处理，不额外收窄结果；带前后空白时会先裁剪再参与筛选
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 创建笔记
      description: 需要当前用户身份；创建后笔记归属当前请求用户；`status` 未传、为 `null`、空字符串或仅空白时默认保存为 `DRAFT`，带前后空白时会先裁剪再保存；`visibility` 为 `null`、空字符串或仅空白时按未传处理，最终存为 `null`；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/resumes:
    get:
      summary: 获取我的简历分页列表
      description: 当前实现按请求用户返回数据，联调可通过 `X-User-Key` 指定用户；首次访问会自动为当前用户补建默认档案；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；`status`、`templateKey`、`keyword` 传空字符串或仅空白时按未传处理，带前后空白时会先裁剪再参与筛选；未带身份时返回 `401 NOT_LOGGED_IN`，用户已被封禁时返回 `403 USER_BANNED`
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: status
          in: query
          schema:
            type: string
        - name: templateKey
          in: query
          schema:
            type: string
          description: 按模板筛选
        - name: keyword
          in: query
          schema:
            type: string
          description: 按模板名或文件名模糊筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          description: 业务错误码，便于前端区分类型
        message:
          type: string
          description: 出错提示
    UserProfileDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        name:
          type: string
          nullable: true
        avatarFileId:
          type: string
          format: uuid
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        status:
          type: string
    MeDto:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/UserProfileDto'
        myResourceCount:
          type: integer
          format: int64
        myFavoriteCount:
          type: integer
          format: int64
        myRoadmapCount:
          type: integer
          format: int64
        myNoteCount:
          type: integer
          format: int64
        myResumeCount:
          type: integer
          format: int64
        recentResourceCount:
          type: integer
          format: int64
        publishedResourceCount:
          type: integer
          format: int64
        draftNoteCount:
          type: integer
          format: int64
        generatedResumeCount:
          type: integer
          format: int64
    MeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/MeDto'
        message:
          type: string
...
    ResumePageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              description: 总记录数
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/ResumeDto'
        message:
          type: string
...
    ResumeDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        templateKey:
          type: string
        status:
          type: string
          description: 当前生成状态，如 GENERATED
        fileId:
          type: string
          format: uuid
        fileUrl:
          type: string
          description: 实际返回下载链接 `/api/resumes/{id}/download`
        fileName:
          type: string
          nullable: true
          description: 关联文件记录存在时返回文件名；若底层文件记录已丢失则为 `null`
        fileSize:
          type: integer
          format: int64
          nullable: true
          description: 关联文件记录存在时返回文件大小；若底层文件记录已丢失则为 `null`
        fileCreatedAt:
          type: string
          format: date-time
          nullable: true
          description: 关联文件记录存在时返回文件创建时间；若底层文件记录已丢失则为 `null`
        fileUpdatedAt:
          type: string
          format: date-time
          nullable: true
          description: 关联文件记录存在时返回文件更新时间；若底层文件记录已丢失则为 `null`
    ResourceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        type:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        tags:
          type: array
          nullable: true
          items:
            type: string
        externalUrl:
          type: string
          nullable: true
        objectKey:
          type: string
          nullable: true
        visibility:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        author:
          type: string
          nullable: true
        likes:
          type: integer
          format: int64
        favorites:
          type: integer
          format: int64
        downloadCount:
          type: integer
          format: int64
    ResourcePageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              format: int64
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/ResourceDto'
        message:
          type: string
    NoteDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        contentMd:
          type: string
        visibility:
          type: string
          nullable: true
          description: 可见性；传空字符串或仅空白时最终存为 null
        status:
          type: string
          nullable: true
          description: 笔记状态；创建时未传或空白会回落为 DRAFT
      required:
        - id
        - title
        - contentMd
    NoteUpsertRequest:
      type: object
      properties:
        title:
          type: string
        contentMd:
          type: string
        visibility:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
      required:
        - title
        - contentMd
    NoteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/NoteDto'
        message:
          type: string
    NotePageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              format: int64
            page:
              type: integer
            pageSize:
              type: integer
              items:
                type: array
                items:
                  $ref: '#/components/schemas/NoteDto'
        message:
          type: string
    RelatedNoteDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        ownerKey:
          type: string
          nullable: true
        favorites:
          type: integer
          format: int64
        favorited:
          type: boolean
    RelatedNotePageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              format: int64
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/RelatedNoteDto'
        message:
          type: string
...
    RoadmapWorkbenchDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
          nullable: true
        visibility:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        nodeCount:
          type: integer
        completedNodeCount:
          type: integer
        progressPercent:
          type: integer
    RoadmapWorkbenchPageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              format: int64
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/RoadmapWorkbenchDto'
        message:
          type: string
...